Types and Functions

Index

ACE.LinearACEModel

ACE.PIBasis

ACE.Rn1pBasis

ACE.Scal1pBasis

ACE.State

ACE.SymmetricBasis

ACE.Ylm1pBasis

Types

OneParticleBasis : part of ACEbase, not sure why; discuss.

Missing docstring.

Missing docstring for XState. Check Documenter's build log for details.

ACE.State — Type

struct State the main type for states of input variables (particles). This type is intended only for storing of information but no arithmetic should be performed on it. For the latter, we have the DState.

Missing docstring.

Missing docstring for DState. Check Documenter's build log for details.

Missing docstring.

Missing docstring for ACEConfig. Check Documenter's build log for details.

ACE.PIBasis — Type

mutable struct PIBasis: implementation of a permutation-invariant basis based on the density projection trick.

The standard constructor is

PIBasis(basis1p, N, D, maxdeg)

basis1p : a one-particle basis

N : maximum interaction order

D : an abstract degee specification, e.g., SparsePSHDegree

maxdeg : the maximum polynomial degree as measured by D

ACE.SymmetricBasis — Type

struct SymmetricBasis

Constructors

Option 1: pass a PIBasis

SymmetricBasis(pibasis, φ)

All possible permutation-invariant basis functions will be symmetrised and then reduced to a basis (rather than spanning set)

Option 2: pass a OneParticleBasis

SymmetricBasis(φ, basis1p, maxν, maxdeg;
               Deg = NaiveTotalDegree())

will first construct a PIBasis from these inputs and then call the first constructor.

ACE.LinearACEModel — Type

struct LinearACEModel: linear model for symmetric properties in terms of a SymmetricBasis.

ACE.Rn1pBasis — Type

struct Rn1pBasis <: OneParticleBasis

One-particle basis of the form $R_n(r_{ij})$, i.e., no dependence on species or on $l$.

This does two things: (1) translates the TransformedPolys into a valid one-particle basis; and (2) treat it as having vectorial input, i.e. value is scalar but gradient is vectorial.

The default symbols are :rr for the state and :n for the index of the basis function.

ACE.Ylm1pBasis — Type

struct Ylm1pBasis <: OneParticleBasis

One-particle basis of the form

\[\phi_{lm}({\bm r}) = Y_l^m(\hat{\br r})\]

Fundamental building block of ACE basis sets of the form

\[ R_{nl}^{\mu_i, \mu_j}(r_{ij}) Y_l^m(\hat{\bm r})\]

This type basically just translates the SHBasis into a valid one-particle basis.

ACE.Scal1pBasis — Type

struct Scal1pBasis <: OneParticleBasis

One-particle basis of the form $P_n(x_i)$ for a general scalar, invariant input x. This type basically just translates the TransformedPolys into a valid one-particle basis.

Functions

Deeper Internals